+++ src/control_plane/project_infrastructure/mod.rs
@@ -1,4 +1,5 @@
 mod plan_meilisearch_project_resources;
+mod plan_opensearch_project_resources;
 mod plan_soketi_project_resources;
 mod plan_typesense_project_resources;
 mod prepare_project_services;
@@ -6,6 +7,9 @@ mod prepared_project_service;
 mod project_service_preparation_error;
 
 pub(crate) use plan_meilisearch_project_resources::plan_meilisearch_project_resources;
+pub(crate) use plan_opensearch_project_resources::{
+    opensearch_initial_admin_password, plan_opensearch_project_resources,
+};
 pub(crate) use plan_soketi_project_resources::plan_soketi_project_resources;
 pub(crate) use plan_typesense_project_resources::plan_typesense_project_resources;
 pub(crate) use prepare_project_services::prepare_project_services;